+++ src/test/java/pages/components/RegistrationResultsModal.java
@@ -1,0 +1,4 @@
+package pages.components;
+
+public class RegistrationResultsModal {
+}
