+++ src/test/java/tests/TestBase.java
@@ -1,0 +1,16 @@
+package tests;
+
+import com.codeborne.selenide.Configuration;
+import org.junit.jupiter.api.BeforeAll;
+import pages.RegistrationPage;
+
+public class TestBase {
+    RegistrationPage registrationPage = new RegistrationPage();
+
+    @BeforeAll
+    static void beForeAll() {
+//        Configuration.holdBrowserOpen = true;
+//        Configuration.browserSize = "1920x1080";
+        Configuration.baseUrl = "https://demoqa.com";
+    }
+}
